Pick

+++ Tango/flashDevice.py
@@ -61,6 +61,10 @@ if argParser.version_info:
   print version
   quit()
 
+# Variables for use with Pick.
+title = "Are all devices ready to flash?"
+options = ['yes', 'no']
+
 # This makes it possible to run the script on a Mac the same as on Linux.
 whatami = platform.system()
 
@@ -220,25 +224,27 @@ def flashDevices(userBSP, device):
     userBSP: Path to the bsp image.
   """
   if argParser.flash_device:
-    print "Flashing %s ..." % device
-    subprocess.check_call(["adb", "-s", device, "reboot", "bootloader"])
-    time.sleep(1)
-    subprocess.check_call(["fastboot", "-s", device, "flash",
-                "bootloader", userBSP + "/bootloader.bin"])
-    subprocess.check_call(["fastboot", "-s", device, "flash", "dtb",
-                userBSP + "/tegra124-ardbeg.dtb"])
-    subprocess.check_call(["fastboot", "-s", device, "flash", "boot",
-                userBSP + "/boot.img"])
-    subprocess.check_call(["fastboot", "-s", device, "flash", "system",
-                userBSP + "/system.img"])
-    subprocess.check_call(["fastboot", "-s", device, "flash", "recovery",
-                userBSP + "/recovery.img"])
-    subprocess.check_call(["fastboot", "-s", device, "-w"])
-    subprocess.check_call(["fastboot", "-s", device, "reboot"])
-    print "~"*25 + "Flash Finished for device " + device + "~"*25
-    print "Device will now reboot. This takes about 4 minutes."
-    if device == lastDevice:
-      countdown(245)
+    nextStep = pick(options, title)
+    if nextStep == 0:
+      print "Flashing %s ..." % device
+      subprocess.check_call(["adb", "-s", device, "reboot", "bootloader"])
+      time.sleep(1)
+      subprocess.check_call(["fastboot", "-s", device, "flash",
+                  "bootloader", userBSP + "/bootloader.bin"])
+      subprocess.check_call(["fastboot", "-s", device, "flash", "dtb",
+                  userBSP + "/tegra124-ardbeg.dtb"])
+      subprocess.check_call(["fastboot", "-s", device, "flash", "boot",
+                  userBSP + "/boot.img"])
+      subprocess.check_call(["fastboot", "-s", device, "flash", "system",
+                  userBSP + "/system.img"])
+      subprocess.check_call(["fastboot", "-s", device, "flash", "recovery",
+                  userBSP + "/recovery.img"])
+      subprocess.check_call(["fastboot", "-s", device, "-w"])
+      subprocess.check_call(["fastboot", "-s", device, "reboot"])
+      print "~"*25 + "Flash Finished for device " + device + "~"*25
+      print "Device will now reboot. This takes about 4 minutes."
+      if device == lastDevice:
+        countdown(245)
 
 def nvFlash(device):
   subprocess.check_call(["bash", nvPath + "flash.sh"])
@@ -323,18 +329,14 @@ def main(devices=devices):
   if argParser.unlock_device:
     for device in devices:
       unlock(device)
-  title = "Are all devices ready to flash?"
-  options = ['yes', 'no']
-  nextStep = pick(options, title)
-  if nextStep == 0:
-    if argParser.nv_flash_device:
-      for device in devices:
-        nvFlash(device)
+  if argParser.nv_flash_device:
     for device in devices:
-      flashDevices(bspPath + chrono(), device)
-    for device in devices:
-      installApks(appDatePath, device, appUnzipPath)
-    cleanup()
+      nvFlash(device)
+  for device in devices:
+    flashDevices(bspPath + chrono(), device)
+  for device in devices:
+    installApks(appDatePath, device, appUnzipPath)
+  cleanup()
 
 
 if __name__ == "__main__":
